requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Flask-Limiter
PyPDF2
qrcode
reportlab
Pillow
PyMuPDF
flask-cors